create a module
