+++ cmd/ethrex/ethrex.rs
@@ -9,11 +9,9 @@ async fn main() {
     let matches = cli::cli().get_matches();
     cfg_if::cfg_if! {
         if #[cfg(feature = "l2")] {
-            use launch::launch_l2;
-            launch_l2(matches).await;
+            launch::l2::launch(matches).await;
         } else {
-            use launch::launch_l1;
-            launch_l1(matches).await;
+            launch::l1::launch(matches).await;
         }
     }
 }
